+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/infer/Inferrer.java
@@ -1,0 +1,17 @@
+package io.dockstore.webservice.helpers.infer;
+
+import io.dockstore.common.DescriptorLanguage;
+import io.dockstore.common.EntryType;
+import io.dockstore.webservice.helpers.FileTree;
+import java.util.List;
+
+public interface Inferrer {
+
+    List<Entry> infer(FileTree fileTree);
+
+    public record Entry(EntryType type, DescriptorLanguage language, String path, String name) {
+        public Entry changeName(String newName) {
+            return new Entry(type, language, path, newName);
+        }
+    }
+}
